Simplify block allocation and calc log_2vMm on device
- Block allocation is now trivial, so prepare_blocks() is no longer needed.
- Since iso is copied to device, log_2vMm can be calculated on device.
- Reduced iso from int32 to int8.

+++ radis/lbl/factory.py
@@ -943,7 +943,7 @@ class SpectrumFactory(BandFactory):
 
         # load the data
         df = self.df0
-        iso = df["iso"].to_numpy(dtype=np.int32)
+        iso = df["iso"].to_numpy(dtype=np.uint8)
         v0 = df["wav"].to_numpy(dtype=np.float32)
         da = df["Pshft"].to_numpy(dtype=np.float32)
         El = df["El"].to_numpy(dtype=np.float32)
